# HDDS-13402. Speed up TestOmSnapshotFileSystem