--- a/hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/om/snapshot/TestOmSnapshotFileSystem.java
+++ b/hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/om/snapshot/TestOmSnapshotFileSystem.java
@@ -23,8 +23,6 @@ import static org.apache.hadoop.ozone.OzoneConfigKeys.OZONE_FS_ITERATE_BATCH_SIZ
 import static org.apache.hadoop.ozone.OzoneConsts.OM_KEY_PREFIX;
 import static org.apache.hadoop.ozone.OzoneConsts.OZONE_URI_SCHEME;
 import static org.apache.hadoop.ozone.om.OmSnapshotManager.getSnapshotPath;
-import static org.apache.hadoop.ozone.om.helpers.BucketLayout.FILE_SYSTEM_OPTIMIZED;
-import static org.apache.hadoop.ozone.om.helpers.BucketLayout.LEGACY;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -54,7 +52,6 @@ import java.util.TreeSet;
 import java.util.UUID;
 import java.util.concurrent.TimeoutException;
 import org.apache.commons.io.IOUtils;
-import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.hadoop.fs.CommonConfigurationKeysPublic;
 import org.apache.hadoop.fs.FSDataInputStream;
 import org.apache.hadoop.fs.FSDataOutputStream;
@@ -66,27 +63,26 @@ import org.apache.hadoop.fs.ozone.OzoneFileSystem;
 import org.apache.hadoop.fs.permission.FsPermission;
 import org.apache.hadoop.hdds.client.StandaloneReplicationConfig;
 import org.apache.hadoop.hdds.conf.OzoneConfiguration;
-import org.apache.hadoop.ozone.MiniOzoneCluster;
 import org.apache.hadoop.ozone.OzoneConsts;
 import org.apache.hadoop.ozone.TestDataUtil;
-import org.apache.hadoop.ozone.client.BucketArgs;
 import org.apache.hadoop.ozone.client.ObjectStore;
 import org.apache.hadoop.ozone.client.OzoneBucket;
 import org.apache.hadoop.ozone.client.OzoneClient;
 import org.apache.hadoop.ozone.client.OzoneKey;
 import org.apache.hadoop.ozone.client.OzoneSnapshot;
 import org.apache.hadoop.ozone.client.OzoneVolume;
-import org.apache.hadoop.ozone.om.KeyManagerImpl;
-import org.apache.hadoop.ozone.om.OMConfigKeys;
+import org.apache.hadoop.ozone.om.OmConfig;
 import org.apache.hadoop.ozone.om.OmSnapshotManager;
 import org.apache.hadoop.ozone.om.OzoneManager;
 import org.apache.hadoop.ozone.om.exceptions.OMException;
+import org.apache.hadoop.ozone.om.helpers.BucketLayout;
 import org.apache.hadoop.ozone.om.helpers.OmKeyArgs;
 import org.apache.hadoop.ozone.om.helpers.OpenKeySession;
 import org.apache.hadoop.ozone.om.helpers.SnapshotInfo;
 import org.apache.hadoop.ozone.om.protocol.OzoneManagerProtocol;
 import org.apache.hadoop.security.UserGroupInformation;
 import org.apache.ozone.test.GenericTestUtils;
+import org.apache.ozone.test.NonHATests;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
@@ -99,72 +95,53 @@ import org.slf4j.LoggerFactory;
  * Abstract class for OmSnapshot file system tests.
  */
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public abstract class TestOmSnapshotFileSystem {
-  protected static final String VOLUME_NAME =
-      "volume" + RandomStringUtils.secure().nextNumeric(5);
-  protected static final String BUCKET_NAME_FSO =
-      "bucket-fso-" + RandomStringUtils.secure().nextNumeric(5);
-  protected static final String BUCKET_NAME_LEGACY =
-      "bucket-legacy-" + RandomStringUtils.secure().nextNumeric(5);
+public abstract class TestOmSnapshotFileSystem implements NonHATests.TestCase {
 
-  private MiniOzoneCluster cluster = null;
   private OzoneClient client;
   private ObjectStore objectStore;
   private OzoneConfiguration conf;
   private OzoneManagerProtocol writeClient;
   private OzoneManager ozoneManager;
   private String keyPrefix;
-  private final String bucketName;
+  private String volumeName;
+  private String bucketName;
+  private final BucketLayout bucketLayout;
   private final boolean createLinkedBuckets;
   private FileSystem fs;
   private OzoneFileSystem o3fs;
-  private Map<String, String> linkedBucketMaps = new HashMap<>();
+  private final Map<String, String> linkedBucketMaps = new HashMap<>();
+  private OmConfig originalOmConfig;
 
   private static final Logger LOG =
-      LoggerFactory.getLogger(TestOmSnapshot.class);
+      LoggerFactory.getLogger(TestOmSnapshotFileSystem.class);
 
-  public TestOmSnapshotFileSystem(String bucketName, boolean createLinkedBuckets) throws Exception {
-    this.bucketName = bucketName;
+  protected TestOmSnapshotFileSystem(BucketLayout layout, boolean createLinkedBuckets) {
+    this.bucketLayout = layout;
     this.createLinkedBuckets = createLinkedBuckets;
-    init();
-  }
-
-  private void init() throws Exception {
-    conf = new OzoneConfiguration();
-    conf.setBoolean(OMConfigKeys.OZONE_OM_ENABLE_FILESYSTEM_PATHS, true);
-    cluster = MiniOzoneCluster.newBuilder(conf).build();
-    cluster.waitForClusterToBeReady();
-    client = cluster.newClient();
-
-    objectStore = client.getObjectStore();
-    writeClient = objectStore.getClientProxy().getOzoneManagerClient();
-    ozoneManager = cluster.getOzoneManager();
-
-    TestDataUtil.createVolume(client, VOLUME_NAME);
-    OzoneBucket bucket = TestDataUtil.createBucket(client, VOLUME_NAME,
-        new BucketArgs.Builder().setBucketLayout(FILE_SYSTEM_OPTIMIZED).build(),
-        BUCKET_NAME_FSO, createLinkedBuckets);
-    if (createLinkedBuckets) {
-      linkedBucketMaps.put(bucket.getName(), bucket.getSourceBucket());
-    }
-    bucket = TestDataUtil.createBucket(client, VOLUME_NAME,
-        new BucketArgs.Builder().setBucketLayout(LEGACY).build(),
-        BUCKET_NAME_LEGACY, createLinkedBuckets);
-    if (createLinkedBuckets) {
-      linkedBucketMaps.put(bucket.getName(), bucket.getSourceBucket());
-    }
-
-
-
-    // stop the deletion services so that keys can still be read
-    KeyManagerImpl keyManager = (KeyManagerImpl) ozoneManager.getKeyManager();
-    keyManager.stop();
   }
 
   @BeforeAll
   public void setupFsClient() throws IOException {
+    OmConfig omConfig = cluster().getOzoneManager().getConfig();
+    originalOmConfig = omConfig.copy();
+    omConfig.setFileSystemPathEnabled(true);
+
+    client = cluster().newClient();
+
+    objectStore = client.getObjectStore();
+    writeClient = objectStore.getClientProxy().getOzoneManagerClient();
+    ozoneManager = cluster().getOzoneManager();
+
+    OzoneBucket bucket = TestDataUtil.createVolumeAndBucket(client, bucketLayout, createLinkedBuckets);
+    if (createLinkedBuckets) {
+      linkedBucketMaps.put(bucket.getName(), bucket.getSourceBucket());
+    }
+    volumeName = bucket.getVolumeName();
+    bucketName = bucket.getName();
+
     String rootPath = String.format("%s://%s.%s/",
-        OzoneConsts.OZONE_URI_SCHEME, bucketName, VOLUME_NAME);
+        OzoneConsts.OZONE_URI_SCHEME, bucketName, volumeName);
+    conf = new OzoneConfiguration(cluster().getConf());
     // Set the fs.defaultFS and start the filesystem
     conf.set(CommonConfigurationKeysPublic.FS_DEFAULT_NAME_KEY, rootPath);
     // Set the number of keys to be processed during batch operate.
@@ -177,9 +154,7 @@ public abstract class TestOmSnapshotFileSystem {
   void tearDown() {
     IOUtils.closeQuietly(client);
     IOUtils.closeQuietly(fs);
-    if (cluster != null) {
-      cluster.shutdown();
-    }
+    cluster().getOzoneManager().getConfig().setFrom(originalOmConfig);
   }
 
   /**
@@ -198,24 +173,16 @@ public abstract class TestOmSnapshotFileSystem {
     }
 
     for (FileStatus fStatus : fileStatuses) {
+      assertEquals(fs.getScheme(), fStatus.getPath().toUri().getScheme(), "unexpected scheme");
       fs.delete(fStatus.getPath(), true);
     }
-
-
-    GenericTestUtils.waitFor(() -> {
-      try {
-        return fs.listStatus(root).length == 0;
-      } catch (Exception e) {
-        return false;
-      }
-    }, 1000, 120000);
   }
 
   @Test
   // based on TestObjectStoreWithFSO:testListKeysAtDifferentLevels
   public void testListKeysAtDifferentLevels() throws Exception {
-    OzoneVolume ozoneVolume = objectStore.getVolume(VOLUME_NAME);
-    assertEquals(ozoneVolume.getName(), VOLUME_NAME);
+    OzoneVolume ozoneVolume = objectStore.getVolume(volumeName);
+    assertEquals(ozoneVolume.getName(), volumeName);
     OzoneBucket ozoneBucket = ozoneVolume.getBucket(bucketName);
     assertEquals(ozoneBucket.getName(), bucketName);
 
@@ -282,7 +249,7 @@ public abstract class TestOmSnapshotFileSystem {
     deleteSnapshot(snapshotName);
     String expectedMessage = String.format("Unable to load snapshot. " +
             "Snapshot with table key '/%s/%s/%s' is no longer active",
-        VOLUME_NAME, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName);
+        volumeName, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName);
     OMException exception = assertThrows(OMException.class,
         () -> ozoneBucket.listKeys(keyPrefix + "a/", null));
     assertEquals(expectedMessage, exception.getMessage());
@@ -370,7 +337,7 @@ public abstract class TestOmSnapshotFileSystem {
 
     // Read using filesystem.
     String rootPath = String.format("%s://%s.%s/", OZONE_URI_SCHEME,
-        bucketName, VOLUME_NAME);
+        bucketName, volumeName);
     OzoneFileSystem o3fsNew = (OzoneFileSystem) FileSystem
         .get(new URI(rootPath), conf);
     try (InputStream fsDataInputStream = o3fsNew.open(new Path(key))) {
@@ -497,21 +464,21 @@ public abstract class TestOmSnapshotFileSystem {
         () -> fs.listStatus(snapshotRoot1));
     assertEquals(String.format("Unable to load snapshot. " +
             "Snapshot with table key '/%s/%s/%s' is no longer active",
-        VOLUME_NAME, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName1), exception1.getMessage());
+        volumeName, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName1), exception1.getMessage());
 
     deleteSnapshot(snapshotName2);
     FileNotFoundException exception2 = assertThrows(FileNotFoundException.class,
         () -> fs.listStatus(snapshotRoot2));
     assertEquals(String.format("Unable to load snapshot. " +
             "Snapshot with table key '/%s/%s/%s' is no longer active",
-        VOLUME_NAME, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName2), exception2.getMessage());
+        volumeName, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName2), exception2.getMessage());
 
     deleteSnapshot(snapshotName3);
     FileNotFoundException exception3 = assertThrows(FileNotFoundException.class,
         () -> fs.listStatus(snapshotParent3));
     assertEquals(String.format("Unable to load snapshot. " +
             "Snapshot with table key '/%s/%s/%s' is no longer active",
-        VOLUME_NAME, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName3), exception3.getMessage());
+        volumeName, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName3), exception3.getMessage());
   }
 
   @Test
@@ -546,7 +513,7 @@ public abstract class TestOmSnapshotFileSystem {
         () -> fs.listStatus(snapshotParent));
     assertEquals(String.format("Unable to load snapshot. " +
             "Snapshot with table key '/%s/%s/%s' is no longer active",
-        VOLUME_NAME, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName), exception.getMessage());
+        volumeName, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName), exception.getMessage());
   }
 
   @Test
@@ -582,7 +549,7 @@ public abstract class TestOmSnapshotFileSystem {
         () -> fs.listStatus(snapshotParent));
     assertEquals(String.format("Unable to load snapshot. " +
             "Snapshot with table key '/%s/%s/%s' is no longer active",
-        VOLUME_NAME, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName), exception.getMessage());
+        volumeName, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName), exception.getMessage());
   }
 
   @Test
@@ -623,11 +590,11 @@ public abstract class TestOmSnapshotFileSystem {
         () -> fs.open(fileInSnapshot));
     assertEquals(String.format("FILE_NOT_FOUND: Unable to load snapshot. " +
             "Snapshot with table key '/%s/%s/%s' is no longer active",
-        VOLUME_NAME, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName), exception.getMessage());
+        volumeName, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName), exception.getMessage());
   }
 
   private void createAndCommitKey(String keyName) throws IOException {
-    OmKeyArgs keyArgs = new OmKeyArgs.Builder().setVolumeName(VOLUME_NAME)
+    OmKeyArgs keyArgs = new OmKeyArgs.Builder().setVolumeName(volumeName)
         .setBucketName(bucketName).setKeyName(keyName)
         .setAcls(Collections.emptyList())
         .setReplicationConfig(StandaloneReplicationConfig.getInstance(ONE))
@@ -673,7 +640,7 @@ public abstract class TestOmSnapshotFileSystem {
         () -> fs.listStatus(snapshotRoot));
     assertEquals(String.format("Unable to load snapshot. " +
             "Snapshot with table key '/%s/%s/%s' is no longer active",
-        VOLUME_NAME, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName), exception.getMessage());
+        volumeName, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName), exception.getMessage());
   }
 
   /**
@@ -728,17 +695,17 @@ public abstract class TestOmSnapshotFileSystem {
         () -> fs.listStatus(snapshotRoot));
     assertEquals(String.format("Unable to load snapshot. " +
             "Snapshot with table key '/%s/%s/%s' is no longer active",
-        VOLUME_NAME, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName), exception.getMessage());
+        volumeName, linkedBucketMaps.getOrDefault(bucketName, bucketName), snapshotName), exception.getMessage());
   }
 
   private String createSnapshot(String snapshotName)
       throws IOException, InterruptedException, TimeoutException {
 
     // create snapshot
-    writeClient.createSnapshot(VOLUME_NAME, bucketName, snapshotName);
+    writeClient.createSnapshot(volumeName, bucketName, snapshotName);
 
     // wait till the snapshot directory exists
-    OzoneSnapshot snapshot = objectStore.getSnapshotInfo(VOLUME_NAME, bucketName, snapshotName);
+    OzoneSnapshot snapshot = objectStore.getSnapshotInfo(volumeName, bucketName, snapshotName);
     SnapshotInfo snapshotInfo = ozoneManager.getMetadataManager()
         .getSnapshotInfoTable()
         .get(SnapshotInfo.getTableKey(snapshot.getVolumeName(), snapshot.getBucketName(), snapshotName));
@@ -751,6 +718,6 @@ public abstract class TestOmSnapshotFileSystem {
   }
 
   private void deleteSnapshot(String snapshotName) throws IOException {
-    writeClient.deleteSnapshot(VOLUME_NAME, bucketName, snapshotName);
+    writeClient.deleteSnapshot(volumeName, bucketName, snapshotName);
   }
 }