## What changes were proposed in this pull request?

Run `TestOmSnapshotFileSystem*` with a single cluster to reduce setup time.

https://issues.apache.org/jira/browse/HDDS-13402

## How was this patch tested?

Before:

```
[INFO] Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 44.06 s -- in org.apache.hadoop.ozone.om.snapshot.TestOmSnapshotFileSystemLegacy
[INFO] Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 40.22 s -- in org.apache.hadoop.ozone.om.snapshot.TestOmSnapshotFileSystemFsoWithLinkedBuckets
[INFO] Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 43.92 s -- in org.apache.hadoop.ozone.om.snapshot.TestOmSnapshotFileSystemLegacyWithLinkedBuckets
[INFO] Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 40.47 s -- in org.apache.hadoop.ozone.om.snapshot.TestOmSnapshotFileSystemFso
```

After:

```
[INFO] Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 22.89 s -- in org.apache.hadoop.ozone.om.snapshot.SnapshotTests$OmSnapshotFileSystemLegacyWithLinkedBuckets
[INFO] Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 18.40 s -- in org.apache.hadoop.ozone.om.snapshot.SnapshotTests$OmSnapshotFileSystemLegacy
[INFO] Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 15.18 s -- in org.apache.hadoop.ozone.om.snapshot.SnapshotTests$OmSnapshotFileSystemFsoWithLinkedBuckets
[INFO] Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 14.30 s -- in org.apache.hadoop.ozone.om.snapshot.SnapshotTests$OmSnapshotFileSystemFso
[INFO] Tests run: 0, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 92.50 s -- in org.apache.hadoop.ozone.om.snapshot.TestOzoneSnapshotsNonHA
```

Total 92.5 seconds instead of 168.67.

https://github.com/adoroszlai/ozone/actions/runs/16119193466